// angular/src/app/shared/modules/patient-contacts/patient-contacts.enum.ts
export enum PatientContactsTypes {
    HOME_ID = 1,
    WORK_ID = 2,
    CELL_ID = 3,
    OTHER_ID = 4,
    EMAIL_ID = 5,
    FAX_ID = 6,
}

export enum OrganizationContactsTypes {
    EMAIL_ID = 1,
    FAX_ID = 2,
    PHONE_ID = 8,
    WEBSITE_ID = 16,
}

export enum SourceContactTypes {
    Patient = 'patient',
    Organization = 'organization',
}